# resnext101_32x4d oriented optimized implementations for packed memory.