--- a/.circleci/unittest/linux/scripts/install.sh
+++ b/.circleci/unittest/linux/scripts/install.sh
@@ -40,22 +40,20 @@ fi
 
 NIGHTLY_DATE=20210606
 
-# if [ "${CU_VERSION:-}" == cpu ] ; then
-#     pip install -q https://download.pytorch.org/whl/nightly/cpu/torch-1.10.0.dev${NIGHTLY_DATE}%2Bcpu-${PYVSHORT}-linux_x86_64.whl
-#     pip install -q https://download.pytorch.org/whl/nightly/cpu/torchvision-0.11.0.dev${NIGHTLY_DATE}%2Bcpu-${PYVSHORT}-linux_x86_64.whl
-#     conda install -y ninja
-#     PYTORCH_VERSION="$(python -c "import torch; print(torch.__version__)")" USE_NINJA=1 python setup.py develop bdist_wheel -d $WHEELS_FOLDER
-# else
-#     pip install -q https://download.pytorch.org/whl/nightly/cu102/torch-1.10.0.dev${NIGHTLY_DATE}%2Bcu102-${PYVSHORT}-linux_x86_64.whl
-#     pip install -q https://download.pytorch.org/whl/nightly/cu102/torchvision-0.11.0.dev${NIGHTLY_DATE}-${PYVSHORT}-linux_x86_64.whl
-#     conda install -y ninja
-#     PYTORCH_VERSION="$(python -c "import torch; print(torch.__version__)")" FORCE_CUDA=1 USE_NINJA=1 python setup.py develop bdist_wheel -d $WHEELS_FOLDER
-# fi
-
 if [ "${CU_VERSION:-}" == cpu ] ; then
-    conda install -y pytorch torchvision cpuonly -c pytorch-nightly
+    pip3 install --pre torch==1.9.0dev{NIGHTLY_DATE} torchvision=1.9.0dev{NIGHTLY_DATE} -f https://download.pytorch.org/whl/nightly/cu102/torch_nightly.html
+    conda install -y ninja
     PYTORCH_VERSION="$(python -c "import torch; print(torch.__version__)")" USE_NINJA=1 python setup.py develop bdist_wheel -d $WHEELS_FOLDER
 else
-    conda install -y pytorch torchvision cudatoolkit=10.2 -c pytorch-nightly
+    pip3 install --pre torch=1.9.0dev{NIGHTLY_DATE} torchvision=1.9.0dev{NIGHTLY_DATE} -f https://download.pytorch.org/whl/nightly/cpu/torch_nightly.html
+    conda install -y ninja
     PYTORCH_VERSION="$(python -c "import torch; print(torch.__version__)")" FORCE_CUDA=1 USE_NINJA=1 python setup.py develop bdist_wheel -d $WHEELS_FOLDER
 fi
+
+# if [ "${CU_VERSION:-}" == cpu ] ; then
+#     conda install -y pytorch torchvision cpuonly -c pytorch-nightly
+#     PYTORCH_VERSION="$(python -c "import torch; print(torch.__version__)")" USE_NINJA=1 python setup.py develop bdist_wheel -d $WHEELS_FOLDER
+# else
+#     conda install -y pytorch torchvision cudatoolkit=10.2 -c pytorch-nightly
+#     PYTORCH_VERSION="$(python -c "import torch; print(torch.__version__)")" FORCE_CUDA=1 USE_NINJA=1 python setup.py develop bdist_wheel -d $WHEELS_FOLDER
+# fi