--- a/nestedtensor/csrc/BinaryOps.cpp
+++ b/nestedtensor/csrc/BinaryOps.cpp
@@ -1,4 +1,9 @@
 #include <nestedtensor/csrc/BinaryOps.h>
+#ifdef WITH_CUDA
+#include <c10/cuda/CUDAStream.h>
+#include <nestedtensor/csrc/cuda/add.h>
+#include <c10/util/Half.h>
+#endif
 
 namespace at {
 
@@ -10,8 +15,8 @@ Tensor NestedTensor_add_Tensor(
     const Scalar& alpha) {
   Tensor self = self_;
   Tensor other = other_;
-  std::cout << "add self: " << is_nested_tensor_impl(self) << std::endl;
-  std::cout << "add other: " << is_nested_tensor_impl(other) << std::endl;
+  // std::cout << "add self: " << is_nested_tensor_impl(self) << std::endl;
+  // std::cout << "add other: " << is_nested_tensor_impl(other) << std::endl;
   if (is_nested_tensor_impl(self) && is_nested_tensor_impl(other)) {
     EfficientSizeNode self_efficient_nested_size =
         get_efficient_nested_size(self);
@@ -36,6 +41,79 @@ Tensor NestedTensor_add_Tensor(
     self = NestedTensor_contiguous(self);
     int64_t self_dim = get_dim(self);
     auto self_opt_sizes = get_opt_sizes(self);
+#ifdef WITH_CUDA
+    if (self_dim == 4 && other.dim() == 4 &&
+        self_opt_sizes[0] &&
+        self_opt_sizes[1] &&
+        (*self_opt_sizes[1]) == other.size(1) &&
+        other.size(0) == 1 &&
+        other.size(2) == 1 &&
+        other.size(3) == 1 &&
+        self.dtype() ==  c10::ScalarType::Half &&
+        other.dtype() == c10::ScalarType::Half) {
+      other = other.contiguous();
+      at::Tensor self_buffer = get_buffer(self);
+      // std::cout << "self_buffer.is_contiguous(): " << self_buffer.is_contiguous() << std::endl;
+      // std::cout << "other.is_contiguous(): " << other.is_contiguous() << std::endl;
+      Tensor nt_sizes_ =
+          get_efficient_nested_size(self).sizes().to(torch::kInt32);
+      // std::cout << "nt_sizes_: " << nt_sizes_ << std::endl;
+      Tensor nt_sizes_1 = at::native::narrow(nt_sizes_, 1, 1, 1);
+      Tensor nt_sizes_2 = at::native::narrow(nt_sizes_, 1, 2, 1);
+      // std::cout << "nt_sizes_1: " << nt_sizes_1 << std::endl;
+      // std::cout << "nt_sizes_2: " << nt_sizes_2 << std::endl;
+      // Tensor nt_sizes_all = nt_sizes_0 * nt_sizes_1 * nt_sizes_2;
+      Tensor nt_sizes_all = nt_sizes_1 * nt_sizes_2;
+      std::vector<int> numbers;
+      for (int64_t i = 0; i < nt_sizes_all.size(0); i++) {
+        for (int64_t j = 0; j < *self_opt_sizes[1]; j++) {
+          numbers.push_back(nt_sizes_all[i].item<int>());
+        }
+      }
+      // std::cout << "nt_sizes_all: " << nt_sizes_all << std::endl;
+      at::Tensor numbers_t = torch::tensor(numbers).to(torch::kInt32);
+      // std::cout << "numbers_t: " << numbers_t << std::endl;
+      Tensor nt_sizes_cumsum =
+          at::native::cumsum(numbers_t, 0).to(torch::kInt32).reshape({-1});
+      // std::cout << "nt_sizes_cumsum: " << nt_sizes_cumsum << std::endl;
+      TORCH_CHECK(nt_sizes_.dim() == 2, "NestedTensor metadata of unexpected dimension.")
+      Tensor nt_sizes = at::cat({torch::tensor({0}, torch::kInt32), nt_sizes_cumsum});
+      // std::cout << "nt_sizes: " << nt_sizes << std::endl;
+      nt_sizes = nt_sizes.to(torch::kCUDA);
+      at::cuda::CUDAStream defaultStream = at::cuda::getDefaultCUDAStream();
+      at::Tensor result_buffer = self_buffer.clone();
+
+      // std::cout << "self.dtype(): " << self.dtype() << std::endl;
+      // std::cout << "other.dtype(): " << other.dtype() << std::endl;
+
+      Half* self_ptr = self_buffer.data_ptr<Half>();
+      Half* other_ptr = other.data_ptr<Half>();
+      Half* result_ptr = result_buffer.data_ptr<Half>();
+      // std::cout << "*self_opt_sizes[0]: " << *self_opt_sizes[0] << std::endl;
+      // std::cout << "*self_opt_sizes[1]: " << *self_opt_sizes[1] << std::endl;
+      // std::cout << "other.sizes(): " << other.sizes() << std::endl;
+      nested_tensor::cuda::add_scalar_kernelLauncher(
+          reinterpret_cast<__half*>(self_ptr),
+          reinterpret_cast<__half*>(other_ptr),
+          reinterpret_cast<__half*>(result_ptr),
+          (int)(*self_opt_sizes[0] * *self_opt_sizes[1]),
+          (int)(*self_opt_sizes[0]),
+          nt_sizes.data_ptr<int>(),
+          defaultStream);
+      // std::cout << "result_buffer: " << result_buffer << std::endl;
+      return wrap_buffer(std::move(result_buffer), get_efficient_nested_size(self),
+          get_efficient_nested_stride(self));
+      // std::cout << "nt_sizes: " << nt_sizes << std::endl;
+      // exit(1);
+      // std::cout << "special other.sizes(): " << other.sizes() << std::endl;
+      // std::tie(self, other) = _expand_other_as(self_, other_);
+      // return map_nested_tensor(
+      //     [&alpha](Tensor s, Tensor o) { 
+      //     return at::add(s, o, alpha); },
+      //     self,
+      //     other);
+    }
+#endif
     if (self_opt_sizes[self_dim - 1] && other.dim() == 1 &&
         (*(self_opt_sizes[self_dim - 1])) == other.size(0)) {
       Tensor self_buffer = get_buffer(self);
@@ -51,8 +129,8 @@ Tensor NestedTensor_add_Tensor(
   std::tie(self, other) = _expand_other_as(self_, other_);
   return map_nested_tensor(
       [&alpha](Tensor s, Tensor o) { 
-      std::cout << "s.sizes(): " << s.sizes() << std::endl;
-      std::cout << "o.sizes(): " << o.sizes() << std::endl;
+      // std::cout << "s.sizes(): " << s.sizes() << std::endl;
+      // std::cout << "o.sizes(): " << o.sizes() << std::endl;
       return at::add(s, o, alpha); },
       self,
       other);
@@ -187,11 +265,66 @@ Tensor& NestedTensor_floor_divide_out(
 }
 
 Tensor NestedTensor_mul_Tensor(const Tensor& self_, const Tensor& other_) {
-  Tensor self;
-  Tensor other;
+  Tensor self = self_;
+  Tensor other = other_;
+  if (is_nested_tensor_impl(self) && !is_nested_tensor_impl(other)) {
+    self = NestedTensor_contiguous(self);
+    int64_t self_dim = get_dim(self);
+    auto self_opt_sizes = get_opt_sizes(self);
+#ifdef WITH_CUDA
+    if (self_dim == 4 && other.dim() == 4 &&
+        self_opt_sizes[0] &&
+        self_opt_sizes[1] &&
+        (*self_opt_sizes[1]) == other.size(1) &&
+        other.size(0) == 1 &&
+        other.size(2) == 1 &&
+        other.size(3) == 1 &&
+        self.dtype() ==  c10::ScalarType::Half &&
+        other.dtype() == c10::ScalarType::Half) {
+      other = other.contiguous();
+      at::Tensor self_buffer = get_buffer(self);
+      Tensor nt_sizes_ =
+          get_efficient_nested_size(self).sizes().to(torch::kInt32);
+      Tensor nt_sizes_1 = at::native::narrow(nt_sizes_, 1, 1, 1);
+      Tensor nt_sizes_2 = at::native::narrow(nt_sizes_, 1, 2, 1);
+      Tensor nt_sizes_all = nt_sizes_1 * nt_sizes_2;
+      std::vector<int> numbers;
+      for (int64_t i = 0; i < nt_sizes_all.size(0); i++) {
+        for (int64_t j = 0; j < *self_opt_sizes[1]; j++) {
+          numbers.push_back(nt_sizes_all[i].item<int>());
+        }
+      }
+      at::Tensor numbers_t = torch::tensor(numbers).to(torch::kInt32);
+      Tensor nt_sizes_cumsum =
+          at::native::cumsum(numbers_t, 0).to(torch::kInt32).reshape({-1});
+      TORCH_CHECK(nt_sizes_.dim() == 2, "NestedTensor metadata of unexpected dimension.")
+      Tensor nt_sizes = at::cat({torch::tensor({0}, torch::kInt32), nt_sizes_cumsum});
+      nt_sizes = nt_sizes.to(torch::kCUDA);
+      at::cuda::CUDAStream defaultStream = at::cuda::getDefaultCUDAStream();
+      at::Tensor result_buffer = self_buffer.clone();
+
+      Half* self_ptr = self_buffer.data_ptr<Half>();
+      Half* other_ptr = other.data_ptr<Half>();
+      Half* result_ptr = result_buffer.data_ptr<Half>();
+      nested_tensor::cuda::mul_scalar_kernelLauncher(
+          reinterpret_cast<__half*>(self_ptr),
+          reinterpret_cast<__half*>(other_ptr),
+          reinterpret_cast<__half*>(result_ptr),
+          (int)(*self_opt_sizes[0] * *self_opt_sizes[1]),
+          (int)(*self_opt_sizes[0]),
+          nt_sizes.data_ptr<int>(),
+          defaultStream);
+      return wrap_buffer(std::move(result_buffer), get_efficient_nested_size(self),
+          get_efficient_nested_stride(self));
+    }
+#endif
+  }
   std::tie(self, other) = _expand_other_as(self_, other_);
   return map_nested_tensor(
-      [](Tensor s, Tensor o) { return at::mul(s, o); }, self, other);
+      [](Tensor s, Tensor o) { 
+      std::cout << "mul special s.sizes(): " << s.sizes() << std::endl;
+      std::cout << "mul special o.sizes(): " << o.sizes() << std::endl;
+      return at::mul(s, o); }, self, other);
 }
 
 Tensor& NestedTensor_mul__Tensor(Tensor& self_, const Tensor& other_) {