Includes CUDA kernels for
- add
- mul
- sub
- batchnorm forward

Various metadata improvements by spreading the use of EfficientSizeNode and specializing on nested_dim == 1

Improved treatment of contiguous memory for
- 1x1 conv2d
- relu_

```
model_name:   resnext101_32x4d, bsz: 64, mean±std shapes[2]: 323.17±145.32, mean±std shapes[3]: 363.86±141.50, loop: 1.54s, nt: 0.50s, speedup: 3.09x
model_name:   resnext101_32x4d, bsz: 128, mean±std shapes[2]: 319.36±143.50, mean±std shapes[3]: 361.93±141.42, loop: 3.09s, nt: 0.98s, speedup: 3.14x
model_name:   resnext101_32x4d, bsz: 256, mean±std shapes[2]: 334.34±147.87, mean±std shapes[3]: 365.12±139.76, loop: 6.33s, nt: 2.18s, speedup: 2.91x
```